+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Fixed select peaks tooltip string typo for `HistogramPeakSlider.m`
 - Added `SegmentationPreprocessingTutorial.mlx`
 - Images are default converted to doubles now within `analysis\simpleSegmentation.m`
-- Colormap in `ChoroidApp.m` is now pink
+- Colormap in `ChoroidApp.m` is now pink, added colormap button to `HistogramPeakSlider.m`
 - Closing out of `OCT\crop` without specifying a crop value resets JSON crop value
 
 ### 7Jan2019